将车辆类型字段改为long型

+++ dao/dao-glm-data/src/main/java/com/navinfo/dataservice/dao/glm/model/rd/link/RdLinkLimit.java
@@ -29,13 +29,13 @@ public class RdLinkLimit implements IRow {
 
 	private int linkPid;
 	
-	private int vehicle;
+	private long vehicle;
 	
-	public int getVehicle() {
+	public long getVehicle() {
 		return vehicle;
 	}
 
-	public void setVehicle(int vehicle) {
+	public void setVehicle(long vehicle) {
 		this.vehicle = vehicle;
 	}
